refactor(tests): move call_command fixture to conftest.py for better test organization
